Update download.sh
We don't need to delete the entire `~/.quai` folder. If the target folder for storing `go-quai` data is located in a place associated with the default path with a symbolic link, this script deletes the link and after importing the snapshot, the node starts saving data to the default folder (aka `~/.quai`). My edits only delete the contents of the `~/.quai` folder.
In addition, the snapshot is loaded using `aria2c`, which can reduce the download time in some cases (like at my house, for example) up to 3-4 times.